Issue #1970 - Limit the AVX/AVX2 check to Windows 10 or later.
The system call IsProcessorFeaturePresent() doesn't understand AVX or
AVX2 checks (39 and 40 respectively) on kernels < 10.0.2004, which would
prevent people on Win 7 or 8 with sufficient CPUs from installing the
browser. Unfortunately, this will not check on Win 7 or 8 as a result,
so people may still be ending up with a non-functional browser on very
old hardware. ¯\_(ツ)_/¯